requirements.txt
sentence-transformers>=2.5.1
faiss-cpu>=1.8.0
rank_bm25>=0.2.2
streamlit>=1.32.0
langchain>=0.1.13
openai>=1.14.0
pydantic>=2.6.4
pandas>=2.2.1
tabulate>=0.9.0
